backend/requirements.txt
# --- Web Framework & Server ---
fastapi==0.128.0
uvicorn==0.40.0

# --- Database & ORM ---
SQLAlchemy==2.0.46
psycopg2-binary==2.9.11

# --- Validation & Settings ---
pydantic==2.12.5
pydantic-settings==2.7.1
python-dotenv==1.0.1

# --- Utilities ---
email-validator==2.2.0
requests==2.32.3

# --- Auth & Security ---
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
slowapi==0.1.9

# --- Google Auth ---
google-auth==2.29.0

# --- AI / ML ---
groq==0.9.0
pinecone-client==3.2.2
sentence-transformers==3.0.1
